AB#62111 reduce the size of the CSV
The full report is > 7M and cant be sent via email, in any case >99% of the rows are not needed. The new CSV is ordered, users who have never logged in are removed & the most recent 100 users who have logged in.

Also removed some debugging logging which was left in by mistake